+++ src/platform_impl/windows/window_state.rs
@@ -264,7 +264,22 @@ impl WindowFlags {
         self = self.mask();
         new = new.mask();
 
-        let diff = self ^ new;
+        dbg!(self);
+        dbg!(new);
+
+        let mut diff = self ^ new;
+
+        // when hiding a maximized window, `self` contains `WindowFlags::MAXIMIZED`
+        // but `new` won't have it as it is removed in `new.mask()` call by applying `WindowFlags::INVISIBLE_AND_MASK`
+        // so `diff` will contain `WindowFlags::MAXIMIZED` and that will cause the window to unmaximize, but
+        // since we are trying to hide the window, we need to apply `WindowFlags::INVISIBLE_AND_MASK` on `diff` too.
+        if diff.contains(WindowFlags::MAXIMIZED)
+            && diff.contains(WindowFlags::VISIBLE)
+            && !new.contains(WindowFlags::VISIBLE)
+        {
+            diff &= WindowFlags::INVISIBLE_AND_MASK;
+        }
+
         if diff == WindowFlags::empty() {
             return;
         }